Expressions, some VM work
Finished the expression parsing function. Right now it only supports integers, but I'm trying to devise a way to keep from having to manually write out which instructions to emit.
Also compiled some actual code, manually copied the bytecode to a buffer because I'm a masochist, and ran it through the VM. Found and fixed bugs and repeated and voila, it produces the same arithmetic result as C++ (but much slower!).