Compiler v2
I decided to begin moving the code that works into a new Compiler file. The old one was my first time playing with any sort of compilation, so I learned a lot of lessons and there was a lot of dead code that was not being used. I'm also starting on function parsing, which I'm very pleased with the progress on so far.

Also StringHelpers is an appropriate dumping ground for string query and manipulation functions that really don't belong in the compiler file.